How to rebuild administration in add-new-tab.md (#367)
* How to rebuild administration in add-net-tab.md

I noticed that the only page in the docs that ever mentions that you need to rebuild administration after you change your main.js is add-custom-field.md. This was very confusing to me and I think it should be added to pretty much every page in the docs that makes changes to main.js.