RepositoryIterator: Add note about determinism
See https://github.com/shopware/platform/pull/2070

Since this issue seems to have shown up in the core codebase, I figured it would be good to have a note in the documentation on the importance of having deterministic sorting when using the RepositoryIterator. Otherwise developers risk using the abstraction with a bad sorting which results in unwanted behaviour.

If the issue I'm trying to describe is unclear or you'd like the wording changed please let me know.