Use "tested on" instead of min requirements
Presently, we only show min system requirements in our Documentation. Also in installation steps (when server is already running and configured), only minimum server requirements are checked.

Compatibilities with new PHP versions can only be found in change log when digging for it which is not very beginner friendly. To resolve that, the idea is to show versions that are tested in our CI. Additionally, we propose to report issues with newer versions in our issue tracker.