Add documentation for afterOrderEnabled property
It took me a couple of hours to figure out why custom payment methods disappeared when retrying payments. I don't know if this is the best place to document it, but I couldn't find it anywhere else and since this is the guide most people will use when creating payment methods I think it should be included.